kennel_manager_backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
django-cors-headers>=3.7
whitenoise>=5.3.0
django-environ>=0.8.1
python-dotenv 
six

# Optional packages
django-debug-toolbar>=3.2
django-extensions>=3.1
django-filter>=21.1
